# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/reset-password:
    post:
      summary: 'Reset password.'
      operationId: resetPassword
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: steve93@example.net
                  nullable: false
                token:
                  type: string
                  description: ''
                  example: reprehenderit
                  nullable: false
                password:
                  type: string
                  description: 'Must be at least 6 characters.'
                  example: 'X}8Bz%^*yLF"rUj'
                  nullable: false
              required:
                - email
                - token
                - password
      security: []
...